move profile function to component
